lingbin-temp: temp-commit-at-20241105: Exchange

+++ velox/common/base/RawVector.h
@@ -153,7 +153,7 @@ class raw_vector {
  private:
   // Adds 'bytes' to the address 'pointer'.
   inline T* addBytes(T* pointer, int32_t bytes) {
-    return reinterpret_cast<T*>(reinterpret_cast<uint64_t>(pointer) + bytes);
+    return reinterpret_cast<T*>(reinterpret_cast<uintptr_t>(pointer) + bytes);
   }
 
   // Size with one full width SIMD load worth data above and below, rounded to
@@ -194,12 +194,12 @@ class raw_vector {
   int32_t capacity_{0};
 };
 
-// Returns a pointer to 'size' int32_t's with consecutive values
-// starting at 0. There are at least kPadding / sizeof(int32_t) values
-// past 'size', so that it is safe to access the returned pointer at maximum
-// SIMD width. Typically returns preallocated memory but if this is
-// not large enough,resizes and initializes 'storage' to the requested
-// size and returns storage.data().
+// Returns a pointer to 'size' int32_t's with consecutive values starting at 0.
+// There are at least kPadding / sizeof(int32_t) values past 'size', so that it
+// is safe to access the returned pointer at maximum SIMD width. Typically,
+// returns pre-allocated memory but if this is not large enough, otherwise,
+// resizes and initializes 'storage' to the requested size and returns
+// storage.data().
 const int32_t*
 iota(int32_t size, raw_vector<int32_t>& storage, int32_t offset = 0);
